Edited comments and restructured for demo
Changed the final executable part of the program to run each of the player classes through a game. Edited the comments for a touch more clarity. I also altered the return value of the wordleGameLoop() function to output the score (I'll need that in addition to allowing player objects to reset when I get to comparing different solving strategies in the future).