Added Progress Bar, Targeted Exclusions; Changed Word List
Fixed the bug where Mk II was getting tested again instead of Mk III. 

Added short circuits to the tests in the possibility pruning .

It's debatable which dictionary to use for the Mk. II/III algorithms. For realism we'd want to use words a human would guess, which are all the words in the soln list, but it might be too narrow...
Sure makes it faster though!

Also by making it discard possible words with 'yellow' letters in the same spot that they were in a previous guess, it greatly improved performance for both.